+++ robyn/scaffold/postgres/requirements.txt
@@ -1,2 +1,3 @@
 robyn
-psycopg2
+psycopg2; platform_system=="Windows"
+psycopg2-binary; platform_system!="Windows"
